Fixed logger tests which were failing because of moving the logger to separate package.
